Move Header component to layout folder.
